backend/requirements.txt
fastapi==0.95.2
uvicorn==0.34.0
psycopg2==2.9.3
python-dotenv==1.0.1
pydantic==1.10.21
requests==2.32.3
asyncpg==0.29.0